requirements_serverless.txt
# GodsEye DNA Pipeline - Serverless Requirements

# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.3.0

# Existing pipeline dependencies
PyYAML==6.0.1
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.3.2
openai==1.3.7

# Production server
gunicorn==21.2.0
